refactor(services): remove IdentityService and related dependencies
